+++ Source/Testably.Abstractions/FileSystem.DriveInfoFactory.cs
@@ -1,6 +1,5 @@
 ﻿using System.IO;
 using System.Linq;
-using Testably.Abstractions.Models;
 
 namespace Testably.Abstractions;
