+++ Source/Testably.Abstractions/FileSystem.FileStreamFactory.cs
@@ -1,5 +1,4 @@
 ﻿using System.IO;
-using Testably.Abstractions.Models;
 
 namespace Testably.Abstractions;
