+++ Tests/Testably.Abstractions.Tests/Testably.Abstractions.Tests.csproj
@@ -65,19 +65,19 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="FileSystemDirectoryInfoTests - Copy.Delete.cs">
+    <Compile Update="FileSystemDirectoryInfoTests.Delete.cs">
       <DependentUpon>FileSystemDirectoryInfoTests.cs</DependentUpon>
     </Compile>
-    <Compile Update="FileSystemFileTests - Copy.AppendAllLines.cs">
+    <Compile Update="FileSystemFileTests.AppendAllLines.cs">
       <DependentUpon>FileSystemFileTests.cs</DependentUpon>
     </Compile>
-    <Compile Update="Mock\MockFileSystem - Copy %282%29.FileTests.cs">
+    <Compile Update="Mock\MockFileSystem %282%29.FileTests.cs">
       <DependentUpon>Mock/MockFileSystem.cs</DependentUpon>
     </Compile>
-    <Compile Update="Mock\MockFileSystem - Copy.FileTests.cs">
+    <Compile Update="Mock\MockFileSystem.FileTests.cs">
       <DependentUpon>Mock/MockFileSystem.cs</DependentUpon>
     </Compile>
-    <Compile Update="Parity\ParityTests - Copy.Net6.cs">
+    <Compile Update="Parity\ParityTests.Net6.cs">
       <DependentUpon>Parity/ParityTests.cs</DependentUpon>
     </Compile>
   </ItemGroup>
